+++ previous-versions/banking/src/sdk/accountbalances.ts
@@ -8,6 +8,7 @@ import * as operations from "./models/operations";
 import * as shared from "./models/shared";
 import { SDKConfiguration } from "./sdk";
 import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
+import jp from "jsonpath";
 
 /**
  * Balances for a bank account including end-of-day batch balance or running balances per transaction.
@@ -136,6 +137,20 @@ export class AccountBalances {
                 break;
         }
 
+        res.next = async (): Promise<operations.ListAccountBalancesResponse | null> => {
+            const nextCursor = jp.value(JSON.parse(decodedRes), "");
+            if (nextCursor === undefined) {
+                return null;
+            }
+
+            return await this.list(
+                {
+                    ...req,
+                },
+                retries,
+                config
+            );
+        };
         return res;
     }
 }