+++ previous-versions/banking/src/sdk/models/operations/listbanktransactions.ts
@@ -50,4 +50,6 @@ export class ListBankTransactionsResponse extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     transactions?: shared.Transactions;
+
+    next: () => Promise<ListBankTransactionsResponse | null>;
 }