+++ previous-versions/banking/src/sdk/sdk.ts
@@ -56,8 +56,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.28.0";
-    genVersion = "2.107.3";
+    sdkVersion = "0.28.1";
+    genVersion = "2.108.3";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);