+++ previous-versions/banking/src/sdk/transactioncategories.ts
@@ -8,6 +8,7 @@ import * as operations from "./models/operations";
 import * as shared from "./models/shared";
 import { SDKConfiguration } from "./sdk";
 import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
+import jp from "jsonpath";
 
 /**
  * Hierarchical categories associated with a transaction for greater contextual meaning to transaction activity.
@@ -241,6 +242,20 @@ export class TransactionCategories {
                 break;
         }
 
+        res.next = async (): Promise<operations.ListTransactionCategoriesResponse | null> => {
+            const nextCursor = jp.value(JSON.parse(decodedRes), "");
+            if (nextCursor === undefined) {
+                return null;
+            }
+
+            return await this.list(
+                {
+                    ...req,
+                },
+                retries,
+                config
+            );
+        };
         return res;
     }
 }